refactor, chore: move image decode source to separate source file, add addition notes to comments. closes #18
