+++ src/image/mod.rs
@@ -3,5 +3,6 @@ pub mod backends;
 pub mod image_formats;
 pub mod modifications;
 pub mod image_data;
+pub mod decode;
 
 mod fast_downsample;
